+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#106] Generate valid prefix ingress paths for services using Traefik rewrite middleware.
 
 ## [v6.0.1] - 2026-03-25
 ### Security